scripts/requirements.txt
selenium
requests
psycopg2-binary
python-dotenv
beautifulsoup4